[Devtools] Fix network grouping experiment
There was a regression recently that was fixed (https://codereview.chromium.org/2838673003/) but broke the network grouping experiment. This patch fixes that regression.

We were deeming the item as not needing to be removed or added without checking if there is a new parent.

R=dgozman,pfeldman
BUG=718116

Review-Url: https://codereview.chromium.org/2856863006
Cr-Original-Commit-Position: refs/heads/master@{#469250}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0e9adc6c75e97b47626520733f9a6ac3b3c010ef