Fix crash when fallback canvas surface is too big
It's possible for SkSurface creation of a bitmap to fail for various
reasons, although most commonly because it's too big.

RecordingImageBufferSurface and UnacceleratedImageBufferSurface
should handle the case when the SkSurface for the fallback
surface fails to be created.

Bug: 705498
Change-Id: I348ed2199403053d5690d8c3c0f07a80837f66e3
Reviewed-on: https://chromium-review.googlesource.com/493775
Commit-Queue: enne <enne@chromium.org>
Reviewed-by: Justin Novosad <junov@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#469444}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 541c600149fe0cedcb48649763f137142c467f1f