Render 0px broken images correctly in quirks mode
Even if an image is smaller than the broken image icon it still needs to
render as a replaced element per https://html.spec.whatwg.org/multipage/rendering.html#images-3.
So do that, but ensure the broken image icon is hidden and don't render
it with a border or any other decoration.

BUG=715535

Review-Url: https://codereview.chromium.org/2853733004
Cr-Original-Commit-Position: refs/heads/master@{#469479}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2e3e1972ebbfdcbde70ef7c33c98169641e3f335