Move nested classes in ComputedStyleBase outside.
Currently, groups such as StyleSurroundData are stored as nested classes
in ComputedStyleBase. It is suspected that this is causing an increase
in binary size. In an atttempt to restore the binary size to its
original value, we move the nested classes outside of ComputedStyleBase.
If this succeeds, we will put these classes in a "detail" namespace to
discourage their use externally.

Diff of generated files:
https://gist.github.com/darrnshn/7d475a02af6f5d57b4a1622721111cba/revisions

BUG=718888

Review-Url: https://codereview.chromium.org/2869613002
Cr-Original-Commit-Position: refs/heads/master@{#469911}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a0fc01a3bcb06cac6e10e6eaf2930b1a0fdb73cc