Worklet: Specify JavaScript MIME type for Worklets created from Blob URL
This is a preparation CL to switch Worklet from classic script loading to
module loading.

Worklet script (module script) must be served with one of the valid JavaScript
MIME types[1]. This is required by the HTML spec[2]. This CL changes tests for
PaintWorklet in order to obey the spec.

[1] https://html.spec.whatwg.org/multipage/scripting.html#javascript-mime-type
[2] https://html.spec.whatwg.org/multipage/webappapis.html#fetch-a-single-module-script (step 7)

BUG=627945

Review-Url: https://codereview.chromium.org/2872453002
Cr-Original-Commit-Position: refs/heads/master@{#469912}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9ed576b84a66830d0caf72c4e4823c9b821f6b3f