Generate StyleVisualData in ComputedStyleBase.
This patch generates the following fields as part of a group:
- clip ('storage_only' of type LengthBox)
- HasAutoClip ('storage_only' of type bool)
- text-decoration ('storage_only' of enum type TextDecoration)
- zoom ('storage_only' of type float)

Since HasAutoClip doesn't correspond to a CSS property, we specify
it in ComputedStyleExtraFields.json5.

Furthermore, text-decoration is a keyword CSS property, but it
can take on any subset of values from a set of keywords. We can't
generate this kind of fields yet. Instead of complicating this
patch with the generation of text-decoration, we will specify it as
a storage_only field in ComputedStyleExtraFields.json5 with a
field size of 4 bits (we can't specify it in CSSProperties.json5
since that file doesn't support field sizes).

A future patch will generate text-decoration properly.

Diff of generated files:
https://gist.github.com/b3420eb4c3d0514cd4f1f936181fedb6/revisions

BUG=628043

Review-Url: https://codereview.chromium.org/2841453002
Cr-Original-Commit-Position: refs/heads/master@{#469939}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 60bc5e0281871949b67b8d7d3a2f0bc3db72f98c