Move more users of WebLocalFrameImpl to WebLocalFrameBase.
This CL breaks the dependency of three more classes on WebLocalFrameImpl. To
do this we introduce some pure virtual methods into WebLocalFrameBase that we
override in WebLocalFrameImpl, and then update calls sites to use
WebLocalFrameBase.

In this case the dependencies have been broken for:
- WebAXObject.
- RemoteFrameClientImpl
- ContextMenuClientImpl
- DedicatedWorkerMessagingProxyProviderImpl

BUG=708879

Review-Url: https://codereview.chromium.org/2865953002
Cr-Original-Commit-Position: refs/heads/master@{#469945}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 06c36803ddb5dd35d9baebce676583489cd1d7ed