Upstream service wrkr `importScripts` tests to WPT
In order to contribute this test logic to the Web Platform Tests
project, the invalid assertion it includes needed to be corrected.
Because the test was previously formatted as a single sub-test, the
resultant failure tended to obscure the results of assertions for
independent conditions. Re-factor the test to support more fine-grained
reporting, and white-list the specific failure as "allowed" in the
Chromium build infrastructure.

BUG=688116, 719052
R=mek@chromium.org

Review-Url: https://codereview.chromium.org/2864783002
Cr-Original-Commit-Position: refs/heads/master@{#470105}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e9f34ed893dd0ab15ac4834becee26c4886620d7