Run SharedArrayBuffer WPT tests in virtual/sharedarraybuffer testsuite
This also fixes an issue in postMessage when
CanTransferArrayBuffersAndImageBitmaps is false (e.g. MessagePorts).
Previously, we would remove all transferables from the transferables list,
serialize that, then transfer the transferables separately. This prevents an
exception from being thrown when a SharedArrayBuffer is in the transferables
list.

BUG=chromium:716320

Review-Url: https://codereview.chromium.org/2860303002
Cr-Original-Commit-Position: refs/heads/master@{#470154}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 43a78b926f0d6bdbaadbd804351cac6021415c97