Add CompositionMarkerListImpl
This is the first of the specialized DocumentMarkerList implementations. It is
currently quite similar to GenericDocumentMarkerListImpl, except it doesn't have
the logic for merging Spelling/Grammar markers, and it doesn't implement the
RemoveMarkersUnderWords() method on DocumentMarkerList that will eventually be
moved to SpellCheckMarkerList. Eventually we will stop creating
RenderedDocumentMarkers in non-TextMatchMarkerLists.

BUG=707867

Review-Url: https://codereview.chromium.org/2820343004
Cr-Original-Commit-Position: refs/heads/master@{#470183}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 45abde5e4bc317fc2e91152b1d66feec93673563