Reland explanation: crrev.com/2812893002 caused chrome_public_test_apk failing on chromium.android/Marshmallow. The problem is fixed in this CL.
Note:

To make the payment request tests independent of the network status,
the tests set the time-out = 0,
which means that the sub-keys must never be fetched,
and the address forms will always show a textfield,
and not a dropdown for the admin areas.

Problem with the previous CL:

The function: Handler().PostDelayed receives the time-out,
but may add an additional delay to that.
Therefore, sometimes, even if the time-out is set to 0,
the sub-keys get fetched.

Solution:

To solve this problem, the case for time-out = 0 is checked at first,
and the call to the function PostDelayed is skipped if the condition
holds.

BUG=710911

Review-Url: https://codereview.chromium.org/2866643004
Cr-Original-Commit-Position: refs/heads/master@{#470225}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5e3734d5c1970621238dc4b5fa3d47a8fdd0b090