Remove the parameter of LayoutBox::XXXOverflowRectForPropagation()
The parameter was not documented in the header file, but was required
to be the caller's (parent's) style. Remove the parameter to simplify
callers.

Also extract LayoutBox::RectForOverflowPropagation() for common code.

I created this CL because I need LayoutBox::RectForOverflowPropagation()
in another CL.

R=eae@chromium.org

Review-Url: https://codereview.chromium.org/2863933003 .
Cr-Original-Commit-Position: refs/heads/master@{#470239}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bbfb3316d1d216a98ca569445e304aa6e6f4e6d4