Post task when rejecting Animation promises inside ScriptForbiddenScope
It is possible for the ready promise to be rejected in a ScriptForbiddenScope. Rejecting a promise with a DOMException results in the constructor for DOMException being executed (due to a ToV8 call), which triggers a RELEASE_ASSERT that detects if script is being executed in a forbidden scope. This patch posts a task to reject promises when inside a forbidden scope to prevent the crash.

BUG=701631

Review-Url: https://codereview.chromium.org/2785303002
Cr-Original-Commit-Position: refs/heads/master@{#470391}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 13c178302b8a4ce77dc493127b1a9eff97c4e7c2