Use the default security token for the initial empty Document.
This was incorrectly removed in r412736, but this bug was not noticed
due to an implementation detail of devtools that would trigger the
DidAccessInitialDocument() call via a different code path. When the
devtools changed in r465484, this revealed the bug.

This also improves the tests to try to catch any future instances of
tests that incorrectly pass when they shouldn't by adding baseline
tests that assert that DidAccessInitialDocument() isn't called when
it shouldn't be.

Bug: 718946, 719122
Cq-Include-Trybots: master.tryserver.chromium.linux:linux_site_isolation
Change-Id: I3262f5f4367eeaeae1d24f145136773263610f77
Reviewed-on: https://chromium-review.googlesource.com/502347
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#471040}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6f9445e94f65449f392ea05167979c053b1826cd