Rename IDBRequest::On* callbacks to IDBRequest::Enqueue*.
The current OnSuccess() / OnError() naming scheme is a bit misleading,
because it suggests that the success and error events are dispatched to
script synchronously. The events are actually queued up, via
ExecutionContext::EnqueueEvent(), in order to support having the
application be suspended in Dev Tools.

This CL fixes up the naming scheme and de-virtualizes some IDBRequest
methods that are not overridden by its only subclass, IDBOpenDBRequest.
This leads the way to more complex changes in event processing needed by
the referenced bug.

BUG=713409

Review-Url: https://codereview.chromium.org/2870803002
Cr-Original-Commit-Position: refs/heads/master@{#470844}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 55ad6749e01eb6bb63b6fb2f64655a82cf4e7d2c