Add SpellCheckMarkerListImpl
This marker list implementation can be used for Spelling and Grammar markers.
It maintains the markers in a sorted list and merges markers whose endpoints
touch.

I'm moving RemoveMarkersUnderWords() from DocumentMarkerList to only be on
SpellCheckMarkerListImpl in this CL and adding a cast in
DocumentMarkerController to enable this.

BUG=707867

Review-Url: https://codereview.chromium.org/2829543002
Cr-Original-Commit-Position: refs/heads/master@{#470847}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5107b8dc6ef983872e59f544f14fb68434025cf0