Stop using Resource::ResourceCallback and make Resource have a TaskHandle
Resource::ResourceCallback was made to be static local by crrev.com/16365011.
It was intended to uses a single timer for all Resources.
But I don't think this optimization has an impact in the real world web.
To make the code simple, I'd like to remove the ResourceCallback and make the
Resource have a TaskHandle.

This change will make it easy to support off-main-thread fetch.

BUG=443374

Review-Url: https://codereview.chromium.org/2876453002
Cr-Original-Commit-Position: refs/heads/master@{#470897}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1b18cecc5e51e5946b1c224bcbfd8fd71bd7c23d