Update SameSizeAsComputedStyle to reflect new changes.
SameSizeAsComputedStyle is meant to be tracking the data layout of
ComputedStyle. Recently, we've moved three DataRefs from ComputedStyle
to ComputedStyleBase without updating the size assert. This was a
silent issue because we changed only the layout, not the size.

This patch updates SameSizeAsComputedStyle to reflect the current state
of ComputedStyle.

BUG=628043

Review-Url: https://codereview.chromium.org/2863423002
Cr-Original-Commit-Position: refs/heads/master@{#471092}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 409ff448dc2fdf77c1de19b39cfafd8780066136