Implements CSSPropertyAPI for the d property
A part of Project Ribbon, separating the parsing logic for CSS
properties from the parser into an API. This patch removes
CSSPropertyD from the switch statement in
parseSingleValue, and calls the API instead.

A function pointer to the parseSingleValue function from the API
for the d property is stored in a CSSPropertyDescriptor, and is
called from CSSPropertyParser.

This patch:
- Adds CSSPropertyAPID.cpp to the BUILD.gn file.
- Adds parseSingleValue to "d" in CSSProperties.json5
  so that it will be added to the generated files for the api.
- Moves the parsing logic for d from CSSPropertyParser.cpp to
  CSSPropertyAPID.cpp, which implements
  CSSPropertyAPI.h.

BUG=668012

Review-Url: https://codereview.chromium.org/2642033002
Cr-Original-Commit-Position: refs/heads/master@{#471220}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a90e416a49db962781bf5e3fc5cfd263346e5375