Move more classes to WebLocalFrameBase instead of WebLocalFrameImpl.
This CL uses WebLocalFrameBase to break dependencies on WebLocalFrameImpl at the
web/ layer. In some case, methods have been made pure virtual in WebLocalFrameBase
and overriden in WebLocalFrameImpl to assist with the decoupling.

We need to break these dependencies so we can Move WebLocalFrameImpl to core.

This CL updates:
- TextFinder
- WebEmbeddedWorkerImpl
- WebFrameSerializer
- WebFrameSerializerImpl
- WebSharedWorkerImpl

BUG=708879

Review-Url: https://codereview.chromium.org/2881693002
Cr-Original-Commit-Position: refs/heads/master@{#471308}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ee9b642ef6c3c43f8342cd3522a1367b1a9dcd76