Re-located duplicated service worker test
The Web Platform Tests project maintains a more rigorous version of this
test, but because that version currently fails in Chromium, this version
cannot be removed without negatively effecting test coverage.

Re-name the test file and add an in-line comment to document the test's
deprecated status.

BUG=688116, 507169
R=falken@chromium.org

Review-Url: https://codereview.chromium.org/2882483002
Cr-Original-Commit-Position: refs/heads/master@{#471350}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 036a8648e3fee70b4414d0567fdf278ece77ce8b