Add stable id to PaintImage.
This patch adds a stable id into PaintImage. This is stable in the sense
that if the blink::Image that generated the paint image hasn't changed,
then the id would be the same. It's used in a case where either we have
a partially loaded image that finishes loading or an animated image
ticks to a new frame. In those cases, the SkImage unique id is different
since the contents of the image are different. However, the stable id
should remain the same to indicate that it's a part of logically the
same image.

The primary use for the stable id would be to determine whether the
compositor is allowed to checker the image. Specifically, if we have
ever displayed an image stable id X, then all other images with the
same stable id X are not allowed to be checker imaged.

R=khushalsagar@chromium.org
CQ_INCLUDE_TRYBOTS=master.tryserver.blink:linux_trusty_blink_rel

Review-Url: https://codereview.chromium.org/2873003002
Cr-Original-Commit-Position: refs/heads/master@{#471351}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ae192fde12e7fba4ac2d8f027037d10c3d0f22a4