Fix TalkBack feedback for password fields in Android O
Prior to O, we were supposed to expose either the value of a password field,
or just dots, depending on a value of a system setting, and independent of
what's shown visually.

In O, we're now supposed to expose what's visually shown, whether that's
characters or dots.

HOWEVER, if a character is replaced with a dot we have to update our internal
state, but not fire an Android event with that character change, to keep it
silent.

This fix reverts the change to LayoutText landed by Paul Miller in r468393
(and merged back to M58) because it was suppressing changes to the internal
accessibility state when a character was replaced with a dot. Instead now we
keep the internal state updated but suppress that particular event from being
fired on Android.

As a minor improvement, this change implements setTextSelection
on AccessibilityNodeInfo, which enables TalkBack to announce when
the cursor reaches the beginning or end of a text field the same
way as it does with native fields.

BUG=721663,716212
TBR=wangxianzhu
NOTRY=true

Review-Url: https://codereview.chromium.org/2881723002
Cr-Original-Commit-Position: refs/heads/master@{#471383}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4e36c746c6c51486f9e57e47a946ee35918fd9f0