Remove DocumentMarkerController::MarkersInRange()
Currently, MarkersInRange() is kind of broken in that it can potentially return
DocumentMarkers from multiple Nodes, and it doesn't return any information about
which Node each DocumentMarker is associated with. This means that callers are
currently resorting to dubious methods to attempt to reconstruct the
EphemeralRange associated with the returned marker(s).

After discussing with xiaochengh@ and yosin@, instead of changing the method to
also return the marker's Node or EphemeralRange, we decided to remove the method
and move the logic directly into the callers.

BUG=707867

Review-Url: https://codereview.chromium.org/2857173003
Cr-Original-Commit-Position: refs/heads/master@{#471642}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3639139a6216af6f66840f8348622d45ba06d5b9