Move AppliedTextDecorationList to its own file.
When we generate external types like Length in ComputedStyle, we specify
the path to the type (e.g. "platform/Length"). We then deduce the name
of the type from the path ("Length"). This means that the type must be
stored in a file with the same name.

This patch moves AppliedTextDecorationList to its own file so that it
can be generated.

BUG=628043

Review-Url: https://codereview.chromium.org/2884553002
Review-Url: https://codereview.chromium.org/2884553002
Cr-Original-Commit-Position: refs/heads/master@{#471677}
Committed: https://chromium.googlesource.com/chromium/src/+/762e2d87addf15f32ac5bda17b1ad9a8fb25dee2
Cr-Original-Original-Commit-Position: refs/heads/master@{#471653}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1d5c7897e427cf28b15c6c4aa995f2fe7d16162c