cc: Approximate ideal contents scale under perspective transform.
If we have a perspective transform, we currently fall back to the
device_scale * page_scale as the transform. This yields bad results
when the actual scale applied is significantly different (see
referenced bug). This patch changes that. If we have a perspective
transform, then we map a vector pointing at (1, 1) using the transform
and then use the largest value of x and y (rounded to nearest int) as
an approximate scale. This produces much crisper results in all tests
in the referenced bug. It's not perfect, but perfect is the enemy of
good, right?

R=danakj@chromium.org
BUG=320369
CQ_INCLUDE_TRYBOTS=master.tryserver.blink:linux_trusty_blink_rel

Review-Url: https://codereview.chromium.org/2874913002
Cr-Original-Commit-Position: refs/heads/master@{#471876}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dc5e1afbd1f6b01e44b58ee16f4c556d2ceaa04c